requirements.txt.txt
streamlit
openai-whisper
gtts
eng-to-ipa